+++ VehicleEquipment/Include/SingleAxisPropulsionSystem.hpp
@@ -12,7 +12,7 @@ class SingleAxisPropulsionSystem : public PropulsionSystem
 public:
     SingleAxisPropulsionSystem(Engine&, Engine&, const EngineDriver&);
 
-     void transferCharacteristicValues(const std::pair<int16_t, int16_t>&) override;
+     void transferCharacteristicValues(const std::pair<int32_t, int32_t>&) override;
 
 private:
     Engine& _rightEngine;
