+++ VehicleEquipment/Source/L293DEngineDriver.cpp
@@ -2,7 +2,7 @@
 #include "VehicleConfiguration.hpp"
 
 std::array<std::array<uint8_t, NUMBER_OF_PINS_PER_ENGINE>, NUMBER_OF_ENGINES>
-L293DEngineDriver::calculatePinValues(const std::pair<int16_t, int16_t>& characteristic) const
+L293DEngineDriver::calculatePinValues(const std::pair<int32_t, int32_t>& characteristic) const
 {
     std::array<uint8_t, NUMBER_OF_PINS_PER_ENGINE> pinValuesOfLeftEngine {};
     std::array<uint8_t, NUMBER_OF_PINS_PER_ENGINE> pinValuesOfRightEngine {};
@@ -45,7 +45,7 @@ L293DEngineDriver::calculatePinValues(const std::pair<int16_t, int16_t>& charact
     return {pinValuesOfLeftEngine, pinValuesOfRightEngine};
 }
 
-bool L293DEngineDriver::areCoordinatesInRange(const std::pair<int16_t, int16_t>& coordinates) const
+bool L293DEngineDriver::areCoordinatesInRange(const std::pair<int32_t, int32_t>& coordinates) const
 {
     const auto& [xCoordinate, yCoordinate] = coordinates;
