+++ VehicleEquipment/Test/EngineDriverMock.hpp
@@ -8,6 +8,6 @@ class EngineDriverMock : public EngineDriver
 {
 public:
     MOCK_CONST_METHOD1(calculatePinValues,
-                       std::array<std::array<uint8_t, NUMBER_OF_PINS_PER_ENGINE>, NUMBER_OF_ENGINES>(const std::pair<int16_t, int16_t>&));
+                       std::array<std::array<uint8_t, NUMBER_OF_PINS_PER_ENGINE>, NUMBER_OF_ENGINES>(const std::pair<int32_t, int32_t>&));
 };
